requirements.txt
# gcp
google-adk==1.0.0
google-cloud-secret-manager==2.22.0
